feat: add automated runs for interlay, add retries
