+++ .github/workflows/xcm-interlay-tests.yml
@@ -1,0 +1,42 @@
+---
+name: Daily Interlay XCM tests
+on:
+  schedule:
+    - cron: '5 * * * *' # Run hourly
+  push:
+    branches:
+      - master
+      - test-add-interlay-chopsticks-tests
+
+jobs:
+  chopsticks_test:
+    runs-on: [self-hosted, linux]
+    timeout-minutes: 30
+    container:
+      image: ghcr.io/interlay/chopsticks:master
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v3
+
+      - name: Run Kintsugi tests
+        uses: nick-fields/retry@v2
+        with:
+          max_attempts: 3
+          retry_on: error
+          timeout_minutes: 10
+          command: |
+            cd /app
+            yarn start xcm \
+              --relaychain=configs/polkadot.yml \
+              --parachain=configs/interlay.yml \
+              --parachain=configs/statemint.yml &
+            cd -
+
+            yarn install --frozen-lockfile
+            npx ts-node scripts/interlay-chopsticks-test.ts
+
+      - name: Report status to Discord
+        uses: sarisia/actions-status-discord@v1
+        if: failure()
+        with:
+          webhook: ${{ secrets.DISCORD_WEBHOOK }}
